Bump k8s v0.32 (#451)
* build(deps): bump the k8s group across 1 directory with 3 updates

Bumps the k8s group with 3 updates in the / directory: [k8s.io/apimachinery](https://github.com/kubernetes/apimachinery), [k8s.io/client-go](https://github.com/kubernetes/client-go) and [sigs.k8s.io/kind](https://github.com/kubernetes-sigs/kind).


Updates `k8s.io/apimachinery` from 0.31.3 to 0.32.0
- [Commits](https://github.com/kubernetes/apimachinery/compare/v0.31.3...v0.32.0)

Updates `k8s.io/client-go` from 0.31.3 to 0.32.0
- [Changelog](https://github.com/kubernetes/client-go/blob/master/CHANGELOG.md)
- [Commits](https://github.com/kubernetes/client-go/compare/v0.31.3...v0.32.0)

Updates `sigs.k8s.io/kind` from 0.25.0 to 0.26.0
- [Release notes](https://github.com/kubernetes-sigs/kind/releases)
- [Commits](https://github.com/kubernetes-sigs/kind/compare/v0.25.0...v0.26.0)

---
updated-dependencies:
- dependency-name: k8s.io/apimachinery
  dependency-type: direct:production
  update-type: version-update:semver-minor
  dependency-group: k8s
- dependency-name: k8s.io/client-go
  dependency-type: direct:production
  update-type: version-update:semver-minor
  dependency-group: k8s
- dependency-name: sigs.k8s.io/kind
  dependency-type: direct:production
  update-type: version-update:semver-minor
  dependency-group: k8s
...


* bumb go version in go.mod


* update golang-ci and fix linting issues


* increase lint timeout


---------

+++ pkg/testutils/e2e/checks/checks.go
@@ -53,7 +53,7 @@ type GrpcCheck struct {
 	// Request message
 	Request []byte
 	// Expected return code (default OK)
-	ExpectedStatus int32
+	ExpectedStatus uint32
 	// Delay before attempting access to service
 	Delay time.Duration
 }
@@ -113,8 +113,8 @@ func (c GrpcCheck) Verify(_ kubernetes.Kubernetes, ingress string, namespace str
 		return fmt.Errorf("unexpected error %w", err)
 	}
 
-	if int32(s.Code()) != c.ExpectedStatus {
-		return fmt.Errorf("expected status code %d but %d received", c.ExpectedStatus, int32(s.Code()))
+	if uint32(s.Code()) != c.ExpectedStatus {
+		return fmt.Errorf("expected status code %d but %d received", c.ExpectedStatus, s.Code())
 	}
 
 	return nil
